requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
tk>=0.1.0
requests>=2.30.0 